Add Dummy controller.
In Gnu C, use weak attribute for XY.

This allows the code to link even if XY isn't filled in by the
application, as long as the calling method is not used. The syntax is
GNU-specific.

Add Unix platform.

Add PNG creation and WebSocket server code.

Add WEB chipset, based on PNG Websocket server.

Improve instructions for websocket mode