Fix multi-statement macro
Multi-statement macros can cause evil, evil bugs
that are really hard to track down.  This is because
macros are presumed by most users to effectively
equate to a single statement.

Help meet this expectation, by using the time-honored
technique of wrapping the multiple statements in a
do { ... } while(0) loop, which makes the macro
resolve into a single expression (statement?).