Support WS2812Serial Library on Teensy T4
Recently there was reported that the WS2812Serial library (github.com/PaulStoffregen/WS2812Serial) was not ported over to work on the new Teensy T4, so I thought I would take a look.

I added the T4 support, which is now pending in a Pull Request.  During that I also found that there were errors in one of the core header files for the T4, which I also fixed, which is now pending in another Pull Request (github.com/PaulStoffregen/Cores)

After that was working, it was pointed out that the FastLED sample in the WS2812Serial librry did not compile. So...

More details in the forum Thread:
https://forum.pjrc.com/threads/58442-Non-Blocking-WS2812-LED-Library-and-Teensy-4-0