Improve use of bash printf
Specifically, ensure uncontrolled variables are
first converted to string using bash variable expansion,
and then are printed as a string using "%s", rather than
listing the expanded string directly in the format string.

Similar to C / C++, this avoids the need to escape
special characters, such as the '%' in those variables.