ColorFromPalette by TProgmemRGBGradientPalette
This is a new implementation of a popular ColorFromPalette function. 
It became necessary for me when I began to use gradient palettes from progmem and
also use a structure of variables with implementations of CRGBPalette16 palettes for different effects.
That all takeout an extra 64 bytes of memory for each effect. (((
Now we can use a simply pointer and... It's works! )))

For declaration of the array of palettes:
	...
	const TProgmemRGBGradientPalettePtr gradientPalettes[] = {
		&Sunset_Real_gp[0],	&es_rivendell_15_gp[0], &es_ocean_breeze_036_gp[0],	&rgi_15_gp[0], ...};
	...

To call someone:
	...
	CRGB color = ColorFromPalette(*gradientPalettes[1], index, brightness, blendType);
	...
	fill_palette(leds, NUM_LEDS, startIndex, incIndex, *gradientPalettes[0], brightness, blendType);
	...