Change define NO_CORRECTION in clockless_trinket to NO_CLOCK_CORRECTION (#1263)
The more specific use is in clockless_trinket, so this is the one which has been renamed.
It was only relevant if you also had FASTLED_ALLOW_INTERRUPTS defined.

Existing code will still work exactly the same, but in cases where this define would have altered the code, a warning is now emitted as well.