Add lightness LUT support for AP102
This changes the color intensity to use perceptual lightness when
LED_USE_GLOBAL_BRIGHTNESS=2 is defined.

This increases the dynamic range by orders of magnitude and makes it
much more pleasant at low intensity, even compared to
LED_USE_GLOBAL_BRIGHTNESS=1. The dynamic range is calculated by taking
the per channel intensity into account. This leads to a 256*3*2 bytes
RAM overhead so it cannot be used by default but it's fine for more
powerful devices like a esp8266 or ESP32. I personally tested on a
esp8266 running esphome 2021.11.4 with 150 LEDs. Here's the stats:

  LED_USE_GLOBAL_BRIGHTNESS=0 (default):
    RAM:   [====      ]  38.2% (used 31300 bytes from 81920 bytes)
    Flash: [====      ]  40.6% (used 423688 bytes from 1044464 bytes)

  LED_USE_GLOBAL_BRIGHTNESS=2:
    RAM:   [====      ]  40.1% (used 32836 bytes from 81920 bytes)
    Flash: [====      ]  40.6% (used 424168 bytes from 1044464 bytes)

  RAM:   +1536
  Flash:  +480

Since with this the scaling is done inside the chipset controller, add
loadByte0/1/2() methods to PixelController for gcc.

This is based on the algorithm I wrote for periph.io/x/devices/apa102 a
few years ago. My algorithm there includes temperature in addition to
intensity correction in the ramp. It does provides a slightly better
ramp but doing this here would require a lot more surgery so starting
smaller.

The transitions aren't as nice as they should where the library seems to
be doing a lot of calculation in PixelController that doesn't match to
the resulting lightness generated by the LEDs. Still, the end result
makes a Christmas tree with such LEDs *much more* enjoyable at low
intensity than even LED_USE_GLOBAL_BRIGHTNESS=1.

The algorithm is not enabled for SK9822 since user report was that it
didn't work as expected in practice.

A possible follow up is to use static members for LUT, assuming a use
case of single threaded and multiple APA102Controller instances. I think
it's a niche case so not doing this here.

The LightnessLUT struct assumes memory is zero-initialized.