Update colorutils.h documentation
This was a big one, but all of the color utilities are now thoroughly documented and organized.

There are a few repeated messages in the color palettes classes, as Doxygen was having issues locating symbols for the constructors with array arguments (e.g. `const CRGB rhs[16]`) and the `operator CRGB*` arguments (possibly due to the spaces?). But those are small enough it's not worth worrying about.