Fix Doxygen noise average define in config
If this doesn't exist then there's no issue, but because it *does* for the purposes of document symbol recognition it's causing problems.

Doxygen is complaining that it can't evaluate an expression properly in noise.cpp, as this is compared to an integer and it's defined as blank. To fix the issue this defines it as "0", so the integer comparison works, fails, and everyone is happy. Or at least happy enough to not publish warnings in the document generation log.