Update missing addLeds() documentation blocks
This also includes the addLeds() functions for the third party libraries supported by CFastLED: SmartMatrix, OctoWS2811, and WS2812Serial.

Unfortunately Doxygen has not quite figured out how to handle \copydoc with overloaded functions that have template parameters, so the per-addLeds() comments are limited to the brief. The section header with the parameter names will have to suffice.