Move dithering macros outside of function for doc
This moves the dither macros outside of the PixelController::init_binary_dithering() function so that Doxygen doesn't get confused. This requires duplicating the preprocessor conditional, but since macros aren't scoped to the function regardless I don't see it as an issue.

These probably should be replaced with constants scoped to the function, but that's not a change I want to make right now. Adding a \todo note.

There are no intended functional changes in this commit.