Remove 'struct' from ColorFromPalette arg defs
There was a discrepancy between the argument types for these ColorFromPalette functions in the header and the source. In the header they don't include the 'struct' keyword for the first parameter, and in the source they do. This was causing a function signature mismatch in Doxygen which highlighted the error.

This commit removes the 'struct' keyword so that the source functions match the signature in the header file.

I don't believe this is an actual change to the function signature as the linker has resolved it without issue to date? But I'm adding this change as its own commit because I'm not 100% sure.