Fix Doxygen global brightness define in config
If this doesn't exist then there's no issue, but because it *does* for the purposes of document symbol recognition it's causing problems.

Doxygen is complaining that it can't evaluate an expression properly in chipsets.h, as this is compared to an integer and it's defined as blank. To fix the issue this defines it as "0", so the integer comparison works, fails, and everyone is happy. Or at least happy enough to not publish warnings in the document generation log.

This is the same fix as included in c2fbf05, but with a different config option.