Applies a post gamma correction color scaling.
This preserves color scaling when using setMaxPowerInVoltsAndMilliamps(...)
and setting the global brightness for the strip.

Background:

The assumption with the scaling is that the power ratios between color
components of a given pixel remain the same (within uint8_t precision)
when the scale is applied. Applying the scaling before gamma correction
violates this assumption and this became apparent with field testing. This
cl corrects this by moving the color scaling after the gamma correction has
been applied. The scaling is also done when a given pixel is in their
intermediate 16 bit format, to maximize color accuracy.