Fix Doxygen chipset "unexpected token" warnings
Starting with Doxygen 1.9.5, running Doxygen would result in a host of warnings: "warning: unexpected token TK_COMMAND_BS as the argument of \ifile".

These doxygen warnings were pointing to lines in `chipsets.h`, which was confusing since these seem to be properly documented? It turns out that this was due to the inline comments for the chipsets in `FastLED.h` using copydoc, and referencing their base classes.

This appears to be a bug in Doxygen. Changing these single-line definitions to multi-line (with the associated multi-line comments) fixes the issue and gets rid of the warnings. I'll follow up with a bug report on the Doxygen repo.