scale16by8: explicitly use uint32_t for intermediate computations (#1823)
`i * (1 + scale)` would overflow before the `>> 8` if it was
stored in a uint16_t. Some implicit type conversion must be going
on to make this work, but it's unclear if that implicit type
conversion would reliably work on all platforms. The `1` literal
may be casting `1 + scale` to an `unsigned int`, which on most
platforms is 32 bits, but the C++ standard allows it to be 16 bits.

Fixes https://github.com/FastLED/FastLED/issues/1821