Make RGBWEmulated work with ObjectFLED, ESP3D RMT5, and other controllers (#1871)
* Fix WS2816 to bypass `protected` scope in various platform controller classes

Added a helper class to WS2816Controller that subclasses the platform's
controller and allows access to the controller's protected methods.
Ugly, I'm sorry.

* Fix RGBWEmulated on Teensy4 ObjectFLED and ESP32S3 RMT5

Change the RGBWEmulated controller to call the device
controller's beginShowLeds, show, and endShowLeds methods
in the correct order.  ObjectFLED and RMT5 are the two
controllers that care (that I know of).